+++ lesson-usesubscription-urql/src/components/CommentList.js
@@ -1,6 +1,10 @@
 import React from 'react'
 
 function CommentList({comments = []}) {
+  if (comments.length === 0) {
+    return <div>No Comments</div>
+  }
+
   return (
     <ul>
       {comments.map((comment, index) => (
